Created html template and js to populate html
Sometimes I get carried away and forget to commit. Oops. I iterate through an array of hour objects to create the table, and I use an html template to do so. I also tried starting the time logic but I had to add a military time variable to make it easier.